+++ .gitignore
@@ -8,3 +8,4 @@ dist/
 lib/
 junit.xml
 settings.json
+*.orig
